+++ identity-server/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 # 7.4.0-preview.1
 
+## Enhancements
+- Skip front-channel logout iframe when unnecessary by @bhazen
+
 ## Bug Fixes
 - Reject Pushed Authorization Requests with parameters duplicated in a JAR by @wcabus